pyproject.toml
[project]
name = "fastcampus-aiagent-v1"
version = "1.0.0"
description = "기존 fastcampus-aiagent 실습 코드를 langchain v1로 마이그레이션"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-community",
    "fastapi",
    "neo4j-graphrag>=1.10.0",
    "langgraph>=1.0.0",
    "python-dotenv>=1.0.0",
    "jupyter>=1.1.0",
    "ipykernel>=6.29.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.12.0",
    "uvicorn",
    "pandas",
]